delete encryptFromString, allow decrypt a Charset encoded String
encoding encrypted data via Character set, not Base64 or Hex is not a
common this, and should be avoided for most of the time.
but if we delete MessageProvider#from(String str, Charset encoding), we
cannot encrypt a String.
Making derived inteface of MessageProvider is also an ineffective task.
So I allow users use MessageProvider#from(String str, Charset encoding)
to decrypt from String, even though it may produce a few wrong use cases